Change `add_habit` mutation to take in a `frequency_start_date` instead of a `creation_date` (the habit's first goal will start on this inputted date).
Fix the `add_habit` mutation to account for the `frequency_change_date` of a
`frequency_change_record` having been changed to two fields `start_date` and
`end_date`.

Fix the `add_habit` GraphQL call on the frontend to account for these changes,
and to choose the `frequency_start_date` based on the type of goal (weekly goals
have to start on a Monday).