Fix keyboard shortcut bug (workaround) because keys apparently aren't being tracked properly through keydown and keyup events when multiple are pressed. Specifically keyup it seems. So allow other keys to be "pressed" and part of `keysDown` as long as the shortcut keys are also pressed.
Temp workaround; a flaw of this is that a certain weird combination
of pressing keys simultaneously, including Enter or MetaLeft / Control,
could leave them still part of `addNoteKeysDown` even when not held
down, and just pressing one of them later would trigger the Submit.
Very unlikely the user would ever do this particular pattern, but still
is not ideal to have these issues of tracking pressed keys inaccurately.